refactor: move `case-monitoring` data files to a dedicated folder
Move them to the folder used to store all `case-monitoring` related files.
